+++ sqleto_generator/lib/sqleto_generator.dart
@@ -1,4 +1,4 @@
 library sqleto_generator;
 
-export 'src/sqleto_builder.dart';
-export 'src/sqleto_generator.dart';
+export 'src/schema_builder.dart';
+export 'src/schema_generator.dart';