+++ sqleto_generator/lib/src/sqleto_builder.dart
@@ -1,8 +1,0 @@
-import 'package:build/build.dart';
-import 'package:source_gen/source_gen.dart';
-
-import 'sqleto_generator.dart';
-
-Builder sqletoGenerator(BuilderOptions options) {
-  return SharedPartBuilder([SQLetoGenerator()], 'sqleto_gen');
-}